ref: fix some major issues and format the code

+++ benches/fragmentation.rs
@@ -4,14 +4,14 @@ use std::collections::VecDeque;
 
 fn bench_fragmentation_pattern(c: &mut Criterion) {
     let mut group = c.benchmark_group("fragmentation");
-    
+
     // Alternating allocation/deallocation pattern
     group.bench_function("alternating_alloc_dealloc", |b| {
         let pool = FixedPool::<i32>::new(1000).unwrap();
-        
+
         b.iter(|| {
             let mut handles = Vec::new();
-            
+
             // Allocate every other slot
             for i in 0..500 {
                 handles.push(pool.allocate(i).unwrap());
@@ -19,61 +19,61 @@ fn bench_fragmentation_pattern(c: &mut Criterion) {
                     drop(h); // Immediately deallocate
                 }
             }
-            
+
             // Clean up
             handles.clear();
         });
     });
-    
+
     // FIFO pattern (queue-like)
     group.bench_function("fifo_pattern", |b| {
         let pool = FixedPool::<i32>::new(100).unwrap();
-        
+
         b.iter(|| {
             let mut queue = VecDeque::new();
-            
+
             // Fill queue
             for i in 0..100 {
                 queue.push_back(pool.allocate(i).unwrap());
             }
-            
+
             // Process in FIFO order
             for i in 100..200 {
                 queue.pop_front(); // Deallocate oldest
                 queue.push_back(pool.allocate(i).unwrap());
             }
-            
+
             queue.clear();
         });
     });
-    
+
     // Random access pattern
     group.bench_function("random_lifetime", |b| {
         let pool = FixedPool::<i32>::new(500).unwrap();
-        
+
         b.iter(|| {
             let mut handles = Vec::new();
-            
+
             for i in 0..500 {
-                handles.push(pool.allocate(i).unwrap());
-                
+                handles.push(pool.allocate(i as i32).unwrap());
+
                 // Randomly drop some handles
                 if i % 3 == 0 && !handles.is_empty() {
                     let idx = i % handles.len();
                     handles.remove(idx);
                 }
             }
-            
+
             handles.clear();
         });
     });
-    
+
     group.finish();
 }
 
 fn bench_growing_pool_fragmentation(c: &mut Criterion) {
     let mut group = c.benchmark_group("growing_pool_fragmentation");
-    
+
     group.bench_function("linear_growth", |b| {
         let config = PoolConfig::builder()
             .capacity(100)
@@ -82,26 +82,26 @@ fn bench_growing_pool_fragmentation(c: &mut Criterion) {
             .build()
             .unwrap();
         let pool = GrowingPool::with_config(config).unwrap();
-        
+
         b.iter(|| {
             let mut handles = Vec::new();
-            
+
             // Grow pool through multiple allocations
             for i in 0..500 {
                 if let Ok(h) = pool.allocate(i) {
                     handles.push(h);
                 }
-                
+
                 // Deallocate every 5th
                 if i % 5 == 0 && !handles.is_empty() {
                     handles.remove(0);
                 }
             }
-            
+
             handles.clear();
         });
     });
-    
+
     group.bench_function("exponential_growth", |b| {
         let config = PoolConfig::builder()
             .capacity(10)
@@ -110,37 +110,37 @@ fn bench_growing_pool_fragmentation(c: &mut Criterion) {
             .build()
             .unwrap();
         let pool = GrowingPool::with_config(config).unwrap();
-        
+
         b.iter(|| {
             let mut handles = Vec::new();
-            
+
             for i in 0..500 {
                 if let Ok(h) = pool.allocate(i) {
                     handles.push(h);
                 }
-                
+
                 if i % 7 == 0 && !handles.is_empty() {
                     handles.pop();
                 }
             }
-            
+
             handles.clear();
         });
     });
-    
+
     group.finish();
 }
 
 fn bench_long_running(c: &mut Criterion) {
     let mut group = c.benchmark_group("long_running");
     group.sample_size(10); // Reduce samples for long-running benchmark
-    
+
     group.bench_function("sustained_load", |b| {
         let pool = FixedPool::<i32>::new(1000).unwrap();
-        
+
         b.iter(|| {
             let mut handles = Vec::new();
-            
+
             // Simulate sustained workload
             for cycle in 0..100 {
                 // Allocate batch
@@ -149,18 +149,18 @@ fn bench_long_running(c: &mut Criterion) {
                         handles.push(h);
                     }
                 }
-                
+
                 // Deallocate older batch
                 if handles.len() > 500 {
                     handles.drain(0..100);
                 }
             }
-            
+
             black_box(&handles);
             handles.clear();
         });
     });
-    
+
     group.finish();
 }
